requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
filelock==3.6.0
greenlet==1.1.2
idna==3.3
numpy==1.22.2
pandas==1.4.1
platformdirs==2.5.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
urllib3==1.26.8
virtualenv==20.13.2
